c: language neutral way to specify the kind letter for files included by c preprocessor
In c917dc681cf44ebdbe21bcb9d2400f4fa9e0eb34, I introduced
'I' kind to C parser to record the name of included header files.

For containing more information separators around the file names (<,
", >) are recorded. After thinking this is too aggressive as default
behavior. As other comments, @b4n says "it is odd". I revised the
code in this commit.

In this commit two things are done.
(1) The kind letter for 'I' is changed to 'h' in C language.
(2) Language neutral way to specify the letter for the kind is introduced.
    In older commit, get.c assumes the client was C parser.
    Parsers for other languages cannot define its own kind letter for
    header files. Per parser own control flag for disabling/enablng
    the kind was not avaiable.

@ffes suggested that socpe field can be used for record the differences.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>