read: Allow to unget up to 3 characters
Some parsers need to unget more than one characters, so add support for
this.

Note that this currently breaks some Flex tests because the Flex parser
already is trying to unget several characters, but as this results in
unexpected behavior (only the last ungotten character is read again),
this parser is explicitly broken not to actually care about this.
An upcoming commit will fix this.