Add annotation as a kind apart from interface
The use of default values in annotations was added to Java allow for
syntax such as:

    String message() default "{constraints.fieldmatch}";

This syntax caused ctags to go mad since it doesn't expect to find a
keyword after a closing parenthesis. Which makes sense since it's not
allowed anywhere else in the syntax. Thus, the need for an indepedent
annotation kind.

The new kind enables specific behaviour for this use cause without fear
of breaking parsing for other cases.