Permit a parser having its own letter for file kind
fileKind field is added to parserDefinition type.
It is initialized with KIND_FILE_DEFAULT(F).

Some languages like Ruby uses F for the other
kind already. KIND_FILE_ALT(!) can be used in
such languages. Specify it at the field.

Using the letter specified in fileKind for
the other purpose makes an assertion failure.
It is checked just after initializing the
language.

This is one of the fix for #317.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>