Fortran: parse Fortran 2003 enums
Allow for not-yet-standard enum naming using `:: name` syntax, most of
the code comes from:
https://github.com/geany/geany/commit/9520e7f7d7b7a9570db2feb165653ef1d68e547a

The difference is that in Geany, tags for anonymous enums are also
generated. While in this commit, the anonymous enums is not parsed for
that you can not locate an tag without an unique name. If you want to
tag the enums, you should give them names.