Fortran: don't generate tags for interface content
Import from Geany: https://github.com/geany/geany/commit/329b2cb40135b759322a9a3f942ddad64ae5c78b
Contributed by b4n on 21 Feb 2014

The reason is that the interface just gives a calling procedure the full
knowledge of the types and characteristics of the dummy arguments that
are used inside of the procedure that it references.

However, the content of interface is not the definition of procedures.
When generating tags for source code, what we want is exactly the
definisions.