fortran: fix a memory leakage arising when broken extends statement is given
If an extends statement without a close parenthesis like

   type, extends(shape :: rectangle

is given as input, a memory object assigned to parentType field
in skipStatementIfKeyword will be lost the reference because the
field is over-assigned.

This patch checks the value of the parentType field before
assigning new memory object to it. If a memory object is
already assigned, free it before re-assigning.

See also #341.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>