refactor: remove a file static variable used for calling stringListHasTest
stringListHasTest was taken only two arguments; a stringList itself and
a function for testing. So to find a something in the stringList, you
may have to pass a key for finding via file static variable.

In this commit, an extra parameter is added to stringListHasTest.
The extra parameter is void* and and can be used passing something
you want. As the result you don't have to use the file static variable
for the purpose.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>