Introduce --kinds-<LANG>= option as an alternative to --<LANG>-kinds=
ctags takes options in which <LANG> name is included: --alias-<LANG>,
--corpus-<LANG>, --map-<LANG>, --regex-<LANG>, and --xcmd-<LANG>.

(Some of the added by me to fishman/ctags.)

All they are used for defining or customizing the behavior of
<LANG> parser. These options come from two parts: action and
<LANG>. The actions are preceded <LANG> in all options.

The exception is --<LANG>-kinds. In this commit I introduce
--kinds-<LANG> option as an alternative to --kinds-<LANG>
to make the order consistent with the other options.

--<LANG>-kinds is not removed to keep the compatibility between
exuberant-ctags.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>