Make the hash table used in keywords.c work like a hash table
Problem 1:
There are nearly 1000 keywords from all the used languages and the size
of the table is 128 so there are too many synonyms.

Problem 2:
Even after increasing the number to 2048 the hashing function behaves
poorly causing too many unnecessary synonyms. Using the string hashing
function used in glib improves the situation (also using the prime
modulus).

Problem 3:
Many languages use the same keywords (do, import, try, etc.) and these are
synonyms in the hash table. Incorporate the language to the hashing function
so identical keywords from different languages are placed at different
indexes in the table.

This patch significantly reduces the number of synonyms and the length
of synonym sequences. For the Go parser the lookup function performs twice
as fast (the lookup takes around 15% of the parsing time so the overall
speedup is around 7%).

The speedup however depends parser to parser - e.g. the languages from
c.c avoid the token creation when skipping input so the keyword lookup
happens very rarely and there's no speed difference in this case.