Fortran: don't use trash box for a memory object for signature
After merging 157fd99 valgrind reports memory related trouble.

    $ make units LANGUAGES=Fortran VG=1
    ...
    #valgrind-error:                        6
	    parser-fortran/bug726712.f90
	    parser-fortran/bug734933.f90
	    parser-fortran/debian_432872.f90
	    parser-fortran/fortran-block
	    parser-fortran/fortran-interface
	    parser-fortran/fortran-signature
    ```

The value returned from parseSignature is allocated in  parseSignature.
However, in parseSubprogramFull, the value was copied with vStringNewCopy.
The original value should be used and assigned to signature field of name.

The memory object for signature should be managed carefully because
the memory object gets more references when it
passed to ancestorPush in parseSubprogramFull.

So for managing the memory object for signature, it is better to
use other mechanism than trash box.

The trash box should be used for the memory object only inside
parseSignature to avoid memory leaking when longjmp is called
somewhere in parseSignature and its callees.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>