Fortran: parse Fortran 2003 enums
Allow for not-yet-standard enum naming using `:: name` syntax, see
http://docs.cray.com/books/S-3692-51/html-S-3692-51/z970507905n9123.html

Authored by Colomban Wendling <ban@herbesfolles.org>

The difference from that of Geany is that the kind of enumerator is 'N'
rather than 'F'. This is because KIND_FILE_DEFAULT is 'F' defined in
parse.h. If kind of enumerator is 'F' then doesParserUseKind will return
TRUE which will make an assertion.