main: Fix leaks in parseLongOption() without trashBox
The item value of cookedArgs structure was being used both as a pointer
to existing memory allocations as well as for new memory allocations.
This made it very difficult to cleanly release newly allocated memory
blocks.
As a solution, all assignments to existing memory addresses were
replaced by memory duplications and then a eFree() call was added to
free this new memory allocation when no longer required. This is less
efficient, but the coherency allows for proper memory deallocation.