d: check blockName before building scope string
scope string is built in findScopeHierarchy.  blockName of
statementInfo is referred when building the string.

When input is broken as D syntax, blockName can hold unexpected
value. This was verified by Assert() statement.  Actually d-fha-0 test
case of fuzz-hitting-assertions made the assertion failure.

It is undesirable that broken input makes ctags stop. So in this commit
the assertion if replaced with code that just aborts building the
scope string. In addition lines expecting the scope string are also
skipped.

Removing duplicated vStringClear is suggested by @b4n.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>