refactor, options: make parseOptions file static function
I want to track the place where parseOptions is used.
It will be helpful for debugging optlib.

parseOptions is used both main.c and options.c.
To track the place where parseOptions is used, this commit introduce
parseCmdlineOptions function.

parseCmdlineOptions is extern'ed and can be used outside of options.c.
parseOptions is made file static and can be used inside options.c.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>