javascript: Improve emission of implicit semicolons
Instead of emitting the semicolon right when encountering the newline,
delay this until we find the next token.  This allows to check whether
a semicolon should be inserted between two token, instead of only after
one.

This allows to handle properly a newline in either side of i.e. a
binary operator, as obviously there shouldn't be a semicolon in either
side.

For example, it permits to handle those two cases the same:

	1 +
	2;

	1
	+ 2;

Before, there was no way to know not to insert a semicolon at the end
of the first line in the second case.  Now, tokens on both sides are
checked, and so in this example no extra semicolon is emitted.