options: introduce {optional} long flag for --langdef
You can define a parser but disabled by default with this ``optional``
long flag. Following is an example to disable ``foo`` parser
in definition::

	$ ctags --langdef=foo'{optional}'

This is helpful to add a parser for rather minor language.  This
option is useful in the case you want to add a parser which large parts
of users may will not use. You can enable it with putting
``--languages=+foo`` or ``--languages=all``; and can put it
to your .ctags.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>