javascript: Fix recognizing objects with a trailing comma
If an object ended with an extra comma after the last property it was
incorrectly assumed not to be a class candidate but a mere hash map or
similar.

This lead to recognizing:

	var foo = { a:1 };

and:

	var foo = { a:1, };

differently.

This is silly, and probably not what was originally intended.  I think
the original goal for this code was to bail out of `{}` and not
recognize it as a class, but the code probably evolved gradually to the
current state which makes the extra decision useless, and wrong.