main, INCOMPATIBLE: make adding a pattern to maps with --langmap= exclusive
This commit introduces behavior incompatible with exuberant ctags.

Adding an extension to maps of a language with --langmap=[+]LANG:.EXT
is exclusive; if maps of another lang include .EXT, .EXT is removed
from them when .EXT is added to the maps of LANG.
Only one language in ctags can have .EXT in its maps.

However, Adding a pattern to maps of a language --langmap=[+](PATTERN)
was not exclusive. This commit makes adding a pattern with --langmap
exclusive.

If you want to add an extension and/or a pattern to maps non-exclusively,
use --map-<LANG>= option.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>