units: rewrite option-loading-corpus test case
The original test case targets .m extension.
The extension was used both ObjectiveC and MatLab.
To solve the confliction between the two languages, build-in
tg tables were used. In the test case the build-in table
was replaced with a newly loaded corpus file.

However, we will use newly introduced selector function; and throw the
tables away to solve the confliction. So this test case cannot
use the confliction.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>