units: don't use input file as extra data
Units/option-loading-corpus.d/input.m was used twice in the same test
case; once as a test input file and once as a corpus file.

This becomes an issue to implement "make clean-units" in environment
where builddir and srcdir are separated.

As the test input file, Units test harness can rearranged the path for
input.m properly. There is no issue here.

As the corpus file, input.m was refereed from args.ctags with relative
path. Units test harness doesn't understand the contents of args.ctags.
Units test harness cannot rearranged the path for the corpus file suitably.
So Units test harness will copy the corpus file to the builddir. So
ctags under testing can find the corpus file.

The issue arises when "make clean-units" is run. The make target tries to
delete generated files and keep the original files. As a corpus file
input.m was a generated file, and as a test input file input.m was an
original file. The target cannot decide whether it should delete the
file or not.

So in this commit, I give different file names for the different purposes.

TODO: Write this to docs/units.rst.
Signed-off-by: Masatake YAMATO <yamato@redhat.com>