Fix that writing tags to stdout didn't work with MSYS2
MinGW's mkstemp() deletes a tempfile automatically when the fd is closed.
So we should not close the fd before sorting is finished. Use
redirection when external sort is used.

catFile() is also changed to use file pointer instead of file name. A
file created by mkstemp() might not be able to opened by fopen().