python: Rewrite the parser as a token-based parser
The previous parser reached its limits and was riddled with hacks,
making it very hard, if at all possible, to introduce new features and
fix bugs.
Hopefully a fresh token-based parser should clean everything up and
make future improvements a lot easier.

Note that this commit drops support for Cython, which will be
reintroduced in an upcoming commit.