Introduce --_list-kinds-full option
With --list-kinds option ctags prints a list of of all tag kinds.  An
element in the list includes a kind letter and a kind name or a kind
description.

In the case both kind name and kind description are needed, the
information provided by ctags with the option is not enough.

With newly introduced --_list-kinds-full option, ctags can print the both
field. The output is tab separated and suitable for machine reading.

    $ ./ctags --_list-kinds-full
    Ada	P	packspec	package specifications	on
    Ada	p	package	packages	on
    ...
    Ada	y	annon	loops and blocks with no identifier	off
    Ant	p	project	projects	on
    Ant	P	property	property	on
    Ant	i	import	imports	on
    Ant	t	target	targets	on
    Asm	d	define	defines	on
    ...

As --list-kinds does, the option can take a language name as an option
argument:

	$ ./ctags --_list-kinds-full=Ant
	p	project	projects	on
	P	property	property	on
	i	import	imports	on
	t	target	targets	on

TODO: Xcmd protocol considers only name field of kinds.
      Description field must be handld.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>