C family: reject non-ascii characters as part of identifier
Test case bug1777340.java.t was failed on AppVeyer.
This patch fixes the failure.

The failure comes from the behavior of isalpha
used in isident1 use in parseGeneralToken.
On Fedora 22 it returns 0. On Windows2010 it returns 1. It seems that ctags depends on the
behavior seen on Fedora 22.

    #define isident1(c)  (isalpha(c) || ...

    static void parseGeneralToken (statementInfo *const st, const int c)
    {
	    const tokenInfo *const prev = prevToken (st, 1);

	    if (isident1 (c) || (isLanguage (Lang_java) && isHighChar (c)))

When cppGetc returns STRING_SYMBOL defined in read.h, it is passed to isident1.
STRING_SYMBOL is 338, a value out range of unsigned char.

In other hand man page of isalpha on Fedora says:

    DESCRIPTION
	   These functions check whether c, which must have the value of an unsigned char or EOF...

Therefore we should not pass STRING_SYMBOL to isalpha; a guard condition is needed.

@k-takata suggests me to use isascii for guarding. However, isascii is not enough portable.
So I put the guard condition inline.

See #494, too.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>