Fix a bug for using both regex parser and crafted parser together in a language
Two flags in a parser are used to control how ctags initializes the parser.
Ctags evaluated the flags wrong way. As the result it is impossible
to implement a language with using both regex parser and crafted
parser together in a language.

A regex parser needs a function assigned to parser field of the
parserDefinition. Regex matching are done as side effect of I/O done
by the function. findRegexTags is a specialized function for the
purpose; it just reads the whole contents of given input file.

findRegexTags function is assigned to the field in initialization
phase if the language is implemented only with regex parsers. It
should not be assigned if a crafted parser is also used for
implementing the language; the I/O is done by the crafted parser.

ctags main part(ctags)can know whether findRegexTags should be
assigned to the field or not by combination of following two flags:

With METHOD_NOT_CRAFTED bit flag, a language can let ctags know that
the language is NOT implemented with a crafted parser.

With METHOD_REGEX bit flag, a language can let ctags know that the
part of the language is implemented with a regex parser.

Only when both flags are specified, ctags should assign findRegexTags
function to the parser field of the parserDefinition.

However, there was a bug. Either METHOD_NOT_CRAFTED or METHOD_REGEX
was specified, ctags main part assigned findRegexTags to the field.
This patch fixes this; ctags main part does it only when both
METHOD_NOT_CRAFTED and METHOD_REGEX are specified.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>