Allow regex parser initialize itself before setting up corking
In the future, the cork API will be used to allow regex parser record
scope informations. This is the initial step for the feature.

Whether cork is needed or not in a regex based parser can be
decided after the parser is initialized. Though cork was
initialized before initializing the parser. This patch reorders
the initialization.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>