main: introduce placeholder field to tagEntryInfo
To record anonymous intermediate scope context,
palceholder field of tagEntryInfo and {placeholder} long flag
of regex pattern defintion are introduced.

e.g.

Named scope context:
A {
  B {
    C ...
  }
}

Anonymous scope context:

A {
  {
	C
  }
}

To track the scope context with a stack,
the eanonymous scope context must be recored.
For putting it to stack but not for recording
to tags file, the placeholder field is introduced.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>