lregex: allow built-in regex parsers to use {exclusive} long flag
When {exclusive} long flag is specified, `ignore' field of
associated regexPattern structure should be initialized to TRUE.

However, for built-in regex parser it isn't initialized
because of a bug.

ignore field was initialized in addLanguageRegex. This function
is called only when --regex-<LANG>= is specified. Built-In regex
parser calls addTagRegexInternal via addTagRegex. So the field
is never initialized to TRUE though "{exclusive}" is specified.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>