tmain: handle transformed program name
The program name(ctags by default) is used as prefix in warning
messages.  So comparing expected output and actual output of stderr
can be failed when using a ctags binary is built with
--program-transform-name is used.

To avoid the failures, with this commit tmain target rearranges the
actual output; substitute "^ctags" with the transformed progoram name
before comparing.

The issue is reported by @viccuad on #568.
Close #568.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>