main, json: allow null tag
Close #492.

A null tag was warned and was not outputted.

    $ cat null.json
    [{"":1}]
    $ ./ctags -o - null.json
    ctags: Warning: ignoring null tag in null.json
    0   null.json   /^[{"":1}]$/;"  o

From the view of language spec, it is the right thing that json parser
outputs an empty string though ctags warns it. However, in generally
outputting an empty string as a tag is abnormal.

In this commit, allowNullTag field is added to parserDefinition. If
the field is set to TRUE, a null tag is never warned and is outputted.

Currently only JSON parser set the field to TRUE.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>