main: don't make a tag for an empty string even if AllowNullTag is TRUE
Revisit #492, #590.

As suggested by @b4n in #590, a tag for an empty string is not useful.
With this commit ctags doesn't record such tags even if AllowNullTag is TRUE.

In other word, AllowNullTag field just controls  the warning about
the tags for empty strings; ctags doesn't warn if it is TRUE.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>