Delay calling addTagRegex
Consider the situation you don't use -R option.
The typical case is running "make units".

addTagRegex is called frequently in initialization
stage of parsers. When you don't use -R, ctags
just run parsers of ONE language for reading a given
file.

In this case initializating is needed only for the language.
Initialization for the other languages is waste of time and resource.

This commit delays the calling addTagRegex till really it is needed.
As the result starting up a bit fast.
(--list-kinds and --_list-kinds-full work well.)

Before:

    make units > /dev/null  40.55s user 64.47s system 98% cpu 1:46.67 total
    make units > /dev/null  40.51s user 64.93s system 98% cpu 1:46.82 total

After:

    make units > /dev/null  36.10s user 63.53s system 103% cpu 1:36.70 total
    make units > /dev/null  35.74s user 64.59s system 103% cpu 1:37.08 total

Signed-off-by: Masatake YAMATO <yamato@redhat.com>