Delay installing keywords
Consider the situation you don't use -R option.
The typical case is running "make units".

Many keywords wwere in initialization stage of parsers.

But when you don't use -R, ctags just run parsers of ONE language for
reading a given file. In this case initializating is needed only for
the language.  Initialization for the other languages is waste of time
and resource.

This commit delays the installing keywords till really it is needed.
As the result starting up a bit fast.work well.)

Before:

    make units > /dev/null  36.10s user 63.53s system 103% cpu 1:36.70 total
    make units > /dev/null  35.74s user 64.59s system 103% cpu 1:37.08 total

After:

    make units > /dev/null  34.71s user 64.76s system 103% cpu 1:36.04 total
    make units > /dev/null  34.69s user 64.16s system 103% cpu 1:35.53 total

C and its familiars install keywords a bit complex way. So I don't touch them
in this commit.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>