Introduce 'initialized' flag for managing parser initialization/finalization
'initialize' field was used to track the parser initialization.  The
field was alo used to specify a initializer. Using the same field for
the different purpose is not good style.

This commit introduces a new field 'initialized' for tracking the
parser initialization. Now 'initialize' field is never overwritten
after initialization.

In addition the value 'initialized' is passed to a function specified
with 'finalize' field. So 'finalize' can know whether a function
specified with 'initialize' field is already called or not.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>