Introduce 'tagRegexInstalled' flag for tracking tagRegexTable installation
'tagRegexTable field was used to track the tagRegexTable installation.
The field was also used to specify regex patterns. Using the same field
for the different purpose is not good style.

This commit introduces a new field 'tagRegexInstalled' for tracking the
'tagRegexTable' installation. Now 'tagRegexTable' field is never overwritten
after installation.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>