Introduce 'keywordInstalled' flag for tracking keywordTable installation
'keywordTable field was used to track the keywordTable installation.
The field was also used to specify regex patterns. Using the same field
for the different purpose is not good style.

This commit introduces a new field 'keywordInstalled' for tracking the
'keywordTable' installation. Now 'keywordTable' field is never overwritten
after installation.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>