main: call per parser 'initialize' method after installing tagRegexTable of the parser
This one will be needed in ctagst translator.
To disable a kind for a regex pattern like:

   --langdef=foo
   --regex-foo=/.../s,something/...
   --kinds-foo=-s

A kind 's' of foo is disabled in these options.
's' must be defined before disabling 's'.

The translator translates '--kinds-foo=-s' to a function call; and
puts the call to 'initialize' method of foo.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>