main: simplify the interface between entry.[ch] and field.[ch]
A string being rendered with escaping, a tag and a buffer were passed
to renderEscaped method of fieldDesc as arguments. This commit
simplifies the the interface between entry.[ch] and field.[ch] by
reducing the number of the arguments.

The string is part of the tag so renderEscaped method can choose the
field that should be printed.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>