main: introduce Z field for including "scope:" explicitly as the keyword for scope field
The most of all fields have "key:value" as format, and the "key" is well known.

The exception is scope field. scope field didn't have "scope" as a
key.  Intead the kind name of scope context is used as "key".

This makes a bit difficult a tool reading tags file recognizing a
given field is a scope field or not.

e.g.

	err	parse.c	/^    boolean     err;$/;"	m	struct:getLangCtx	file:

In this line "struct" is the key part of scope field, and "getLangCtx"
is the value part of the field. The key part, "struct", is language specific:

    $ ./ctags --_list-kinds-full=C
    ...
    s	struct	structure names	on

It means the tools must have knowledges about the language and its kinds.

This commit intorudces a well-known key for scope field. With --field=+Z option
"scope" is prefixed to the key of scope.

e.g.

	err	parse.c	/^    boolean     err;$/;"	m	scope:struct:getLangCtx	file:

A tool can use "scope" as a clue for recognizing a scope field in a tag line.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>