fix #629 infinite loop with fortran parser on C file
When parsing C enum, Fortran parse try to find an "end" token in
parseEnumBlock. There is no such key word at end of C enum statement and
the parser falls into infinite loop. To solve it, add an end-of-file
test when trying to find "end" token. If reach end of file, jump out of
the loop.