main: omit patterns for long input lines
Too long input lines may make tags file larger.
This commits omits the length of pattern field printed to a tags file.

If a pattern is longer than the value of Option.patternLengthLimit,
the pattern is omitted; instead of including whole the input line, the
rest of substring after nth byte of Option.patternLengthLimit is
just not printed.

Close #163.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>