C: don't make a tag for switch/case label
A switch case label containing more than two tokens,
C parser mistakenly detects it as a  goto label.

e.g.

	switch (i)
	{
		case X|Y:
		     break;

Here Y was captured as a goto label.

This commit adds an extra condition for avoiding the above case.
Only if a label candidate is not in the switch statement, c parser
makes a "goto label" kind tag for the candidate.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>