main: Manage "compact source line" as newly introduced C field
The format used in -x output is statically defined.
I'm working on --xformat option, which allow a user to define the format dynamically
variety purposes. Each field letter is used in --xformat for controlling whether
ctags prints the field or not.

"compact source line" is one of elements user may want to control its printing.
However it was not managed as a field. For future --xformat implementation,
this commit introduces new field "c,compact" for managing "compact source line".

Signed-off-by: Masatake YAMATO <yamato@redhat.com>