refactor, main: unify duplicated code in writePatternEntry and makePatternString
Generating a pattern is one of the hot-spot in handling very
long input lines. Caching the generated pattern(pattern cache)
for the same input line may reduce the cost.

For implementing the pattern cache, this commit simplifies the
organization of functions(writePatternEntry and makePatternString)
related to Generating patterns.

writePatternEntry and makePatternString does the similar things.
By introducing a function(makePatternStringCommon) doing the things,
the common code can is removed from the original two functions.

The next step for removing the hot-spot is introducing the pattern
cache to the common code, makePatternStringCommon.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>