main: reuse pattern strings
Making a pattern takes long time for a long input line.  This commit
shortens the time. This commit modifies makePatternStringCommon; when
making a pattern for a file position, makePatternStringCommon records
the both file position and the result pattern. If the function is called
again for the same position, reuse the recorded pattern.

Evaluation:

1. Making an input file for evaluation

    $ git clone https://github.com/google/trace-viewer.git
    $ git checkout 2ca1bc034d0055963541aad41ff3ff7fdf483b48
    $ LC_ALL=C ls -lh trace-viewer/tracing/test_data/repaints_on_scroll_region.json
    -rw-r--r--. 1 yamato yamato 9.2M Oct 15 03:08 trace-viewer/tracing/test_data/repaints_on_scroll_region.json

This one is too large for current ctags. Make smaller one.

    $ dd if=trace-viewer/tracing/test_data/repaints_on_scroll_region.json of=tmp-1M.json bs=1024 count=1024
    $ zsh

ctags with this commits is called ctags-with-cache here.
ctags without this commits is called ctags-without-cache here.

    $ time ./ctags-without-cache -o tags-without-cache ./tmp-1M.json
    9.83s user 4.11s system 100% cpu 13.940 total
    $ time ./ctags-with-cache -o tags-with-cache ./tmp-1M.json
    0.12s user 0.07s system 93% cpu 0.204 total
    $ diff tags-without-cache tags-with-cache
    6c6
    < !_TAG_PROGRAM_VERSION	Development	/0ddcb04/
    ---
    > !_TAG_PROGRAM_VERSION	Development	/79fcbd8/

Much faster.

ctags-with-cache passes "make units VG=1".

Signed-off-by: Masatake YAMATO <yamato@redhat.com>