python: check the enablement of "member" kind in makeFunctionTag
In makeFunctionTag function, tags both member and function kinds are
made. If the kinds are disabled, the tags should not be made.

At the head of makeFunctionTag, the enablement of function kind is
checked but that of member is not checked. As the result, an assertion
is made failure as reported in #648 by @yevhen-m.

This commit fixes this; the enablement of "member" kind is also
checked in makeFunctionTag.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>