Python: revise namespace kind and introduce associated roles
This commit introduces incompatible changes about kind names in
python language.

This commit deals with "import ..." and "from X import..." statements.
This commit captures both definition and reference tags from the
statements.

Consider following python code snippet:

	import x
	import x as y
	from x import y as z

Should ctags make a tag on these x, y and z?
This is the fundamental question about this commit.

1. tagging x in `import x'
---------------------------------------------------------------

In this code snippet, `x' defines nothing.
A module name `x' is referenced for importing.
So `x' should be tagged but tagged as a reference tag.

2. tagging x and y in `import x as y'
---------------------------------------------------------------

`x' should be a tagged as a reference tag as written in 1.
`y' defines a new name which refers a module named `x'.
`y' should be tagged as a definition.

3. from x import y as z
---------------------------------------------------------------

`x' is a reference for a module named `x'.
`y' is a reference for a class/variable/function defined in `x'.
`z' defines a name referring something named `y'.

Changed kind names and introduced roles in this commit
---------------------------------------------------------------

Following summarizes what kinds and roles are
used for tagging import statements:

    import X              X = (kind:module, role:imported)

    import X as Y         X = (kind:module, role:indirectly-imported),
			  Y = (kind:namespace, [nameref:X])
			  ------------------------------------------------
			  Don't confuse with namespace role of module kind.

    from X import *       X = (kind:module,  role:namespace)

    from X import Y       X = (kind:module,  role:namespace),
			  Y = (kind:unknown, role:imported, [scope:X])

    from X import Y as Z  X = (kind:module,  role:namespace),
			  Y = (kind:unknown, role:indirectly-imported, [scope:X])
			  Z = (kind:unknown, [nameref:X.Y])

Fields in the brackets are not implemented yet.

Following are --list-* output of ctags with this commit.

    $ ./ctags --_list-kinds-full=Python
    c	class	classes	on	referenceOnly:FALSE	nRoles:0
    f	function	functions	on	referenceOnly:FALSE	nRoles:0
    m	member	class members	on	referenceOnly:FALSE	nRoles:0
    v	variable	variables	on	referenceOnly:FALSE	nRoles:0
    I	namespace	name referring a module defined in other file	on	referenceOnly:FALSE	nRoles:0
    i	module	modules	on	referenceOnly:TRUE	nRoles:3
    x	unknown	name referring a classe/variable/function defined in other module	on	referenceOnly:FALSE	nRoles:2

    $ ./ctags --_list-roles=Python
    Python	i	imported	imported modules	on
    Python	i	namespace	namespace from where classes/variables/functions are imported	on
    Python	i	indirectly-imported	module imported in alternative name	on
    Python	x	imported	imported from the other module	on
    Python	x	indirectly-imported	classes/variables/functions imported in alternative name	on

Comparing the output
---------------------------------------------------------------
Sample input(input.py):

    import a
    import b as c
    from d import e as f

The output of exuberant-ctags:

    $ ctags -o - input.py
    a	input.py	/^import a$/;"	i
    c	input.py	/^import b as c$/;"	i
    f	input.py	/^from d import e as f$/;"	i

`c' and `f' are tagged. These are good. However, `a' is also tagged
though it is not a definition.

The output of universal-ctags with this commit:

    $ ./ctags  -o - foo.py
    c	foo.py	/^import b as c$/;"	I
    f	foo.py	/^from d import e as f$/;"	x

Only `c' and `f' are tagged as definitions.

    $ ./ctags -o - --extra=+r --fields=+r input.py
    a	input.py	/^import a$/;"	i	role:imported
    b	input.py	/^import b as c$/;"	i	role:indirectly-imported
    c	input.py	/^import b as c$/;"	I
    d	input.py	/^from d import e as f$/;"	i	role:namespace
    e	input.py	/^from d import e as f$/;"	x	role:indirectly-imported
    f	input.py	/^from d import e as f$/;"	x

The names of referenced modules in the input are tagged as reference.

Qualified notation also added:

    $ ./ctags -o - --extra=+rq --fields=+r input.py
    a	input.py	/^import a$/;"	i	role:imported
    b	input.py	/^import b as c$/;"	i	role:indirectly-imported
    c	input.py	/^import b as c$/;"	I
    d	input.py	/^from d import e as f$/;"	i	role:namespace
    d.e	input.py	/^from d import e as f$/;"	x	role:indirectly-imported
    e	input.py	/^from d import e as f$/;"	x	role:indirectly-imported
    f	input.py	/^from d import e as f$/;"	x

TODO
---------------------------------------------------------------

A. implement nameref field

For following input, ctags doesn't have proper way to represent the
relation between `b' and `c':

   import b as c

New field like nameref: can be used like:

    c	input.py	/^import b as c$/;"	I	nameref:module:b

B. integrate the change with scope field

For following input, ctags can know `e' is defined in module `d'.

   from d import e

So scope field can be attached to `e` like:

    e	input.py	/^from d import e$/;"	x	role:imported	scope:module:d

Signed-off-by: Masatake YAMATO <yamato@redhat.com>